LoadStoreVectorizer: try harder to identify underlying objects
There is a WMMA kernel where the code for loading 16-bit matrices
requires a higher lookup depth for the underlying object in order to
find it.

This only applies to typed pointers; with opaque pointers, a depth of 4
is sufficient (the default is 6). So this is really just a temporary
workaround not really suitable for upstream.

Change-Id: Idc916035c82416f8c7ce283cbe453c217562cdc7